fix(entity): 修改点击量类型为Long以支持更大数值

+++ shortlink_project/src/main/java/com/cqie/shortlink_project/entity/LinkDO.java
@@ -45,7 +45,7 @@ public class LinkDO extends BaseDO {
     /**
      * 点击量
      */
-    private Integer clickNum;
+    private Long clickNum;
 
     /**
      * 分组标识
